extracted renderer interface
